// src/types/users/students.types.ts
export interface StudentProfile {
    student_id: string
    school: string
    parent_contact_no: string
    barcode: string | null
    user_id: string
    address: string
    admin: any
    avatar: string  | null
    contact_no: string
    dob: string
    email: string
    f_name: string
    l_name: string
    nic: string
    user_type: string
    username: string
  }